move _az_span_token from az_iot_core.h to az_span_internal.h
